move omero metadata code into separate file
